refactor: rename base package to com.example.focustella
