+++ src/test/java/com/example/focustella/FocustellaApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.focustellaapi;
+package com.example.focustella;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
